identity/requirements.txt
wheel
setuptools
twine

mysql-connector-python==8.0.28
jproperties==2.1.1